agente_ia_tech/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1

